Replace generate_def_yaml by a proper object init
