+++ imagecraft/ubuntu_image.py
@@ -18,153 +18,10 @@
 
 import subprocess
 
-from craft_application.util import dump_yaml
 from craft_cli import emit
-from pydantic import BaseModel, Field
 
 from imagecraft.errors import UbuntuImageError
-
-
-def _alias_generator(s: str) -> str:
-    return s.replace("_", "-")
-
-
-class Snap(BaseModel):
-    """Pydantic model for the Snap object in an ImageDefinition."""
-
-    name: str | None
-
-
-class Package(BaseModel):
-    """Pydantic model for the Package object in an ImageDefinition."""
-
-    name: str | None
-
-
-class Customization(BaseModel):
-    """Pydantic model for the Customization object in an ImageDefinition."""
-
-    extra_snaps: list[Snap] | None
-    extra_packages: list[Package] | None
-
-    class Config:
-        """Pydantic model configuration."""
-
-        validate_assignment = True
-        allow_mutation = True
-        allow_population_by_field_name = True
-        alias_generator = _alias_generator
-
-
-class Seed(BaseModel):
-    """Pydantic model for the Seed object in an ImageDefinition."""
-
-    urls: list[str]
-    branch: str
-    names: list[str]
-    pocket: str
-
-    class Config:
-        """Pydantic model configuration."""
-
-        validate_assignment = True
-        allow_mutation = True
-        allow_population_by_field_name = True
-        alias_generator = _alias_generator
-
-
-class Rootfs(BaseModel):
-    """Pydantic model for the Rootfs object in an ImageDefinition."""
-
-    components: list[str]
-    seed: Seed
-
-    class Config:
-        """Pydantic model configuration."""
-
-        validate_assignment = True
-        allow_mutation = True
-        allow_population_by_field_name = True
-        alias_generator = _alias_generator
-
-
-class ImageDefinition(BaseModel):
-    """Pydantic model for the ImageDefinition."""
-
-    name: str
-    display_name: str
-    revision: str
-    class_: str = Field(alias="class")
-    architecture: str
-    series: str
-    kernel: str | None
-    rootfs: Rootfs
-    customization: Customization | None = None
-
-    class Config:
-        """Pydantic model configuration."""
-
-        validate_assignment = True
-        allow_mutation = True
-        allow_population_by_field_name = True
-        alias_generator = _alias_generator
-
-
-def generate_image_def_yaml(  # noqa: PLR0913
-    series: str,
-    revision: str,
-    arch: str,
-    sources: list[str],
-    seed_branch: str,
-    seeds: list[str],
-    components_list: list[str],
-    pocket: str,
-    kernel: str | None = None,
-    extra_snaps: list[str] | None = None,
-    extra_packages: list[str] | None = None,
-) -> str:
-    """Generate a definition yaml file for rootfs creation."""
-    image_definition = ImageDefinition(
-        name="craft-driver",
-        display_name="Craft Driver",
-        class_="preinstalled",  # type: ignore[call-arg]
-        revision=revision,
-        architecture=arch,
-        series=series,
-        kernel=kernel,
-        rootfs=Rootfs(
-            components=components_list,
-            seed=Seed(
-                urls=sources,
-                branch=seed_branch,
-                names=seeds,
-                pocket=pocket,
-            ),
-        ),
-    )
-
-    extra_snaps_obj = None
-    extra_packages_obj = None
-
-    if extra_snaps:
-        extra_snaps_obj = [Snap(name=s) for s in extra_snaps]
-
-    if extra_packages:
-        extra_packages_obj = [Package(name=p) for p in extra_packages]
-
-    if extra_snaps or extra_packages:
-        image_definition.customization = Customization(
-            extra_snaps=extra_snaps_obj,
-            extra_packages=extra_packages_obj,
-        )
-
-    return dump_yaml(
-        image_definition.dict(
-            exclude_unset=True,
-            exclude_none=True,
-            by_alias=True,
-        ),
-    )
+from imagecraft.image_definition import ImageDefinition
 
 
 def ubuntu_image_cmds_build_rootfs(  # noqa: PLR0913
@@ -181,19 +38,21 @@ def ubuntu_image_cmds_build_rootfs(  # noqa: PLR0913
     extra_packages: list[str] | None = None,
 ) -> list[str]:
     """List commands to ubuntu-image to generate a rootfs."""
-    definition_yaml = generate_image_def_yaml(
-        series,
-        version,
-        arch,
-        sources,
-        seed_branch,
-        seeds,
-        components_list,
-        pocket,
-        kernel,
-        extra_snaps,
-        extra_packages,
+    image_def = ImageDefinition(
+        series=series,
+        revision=version,
+        architecture=arch,
+        kernel=kernel,
+        components_list=components_list,
+        seed_urls=sources,
+        seed_branch=seed_branch,
+        seed_names=seeds,
+        seed_pocket=pocket,
+        extra_snaps=extra_snaps,
+        extra_packages=extra_packages,
     )
+
+    definition_yaml = image_def.dump_yaml()
     image_definition_file = "craft.yaml"
 
     return [
