+++ pyproject.toml
@@ -303,6 +303,7 @@ lint.select = [ # Base linting rule selections.
 ]
 lint.ignore = [
     "ANN10", # Type annotations for `self` and `cls`
+    "ANN204",  # Missing type annotations for magic methods
     #"E203",  # Whitespace before ":"  -- Commented because ruff doesn't currently check E203
     "E501",   # Line too long (reason: black will automatically fix this for us)
     "D105",   # Missing docstring in magic method (reason: magic methods already have definitions)
